Fix test of HammingDistance.

+++ pkg/logic/hamming_test.go
@@ -11,14 +11,14 @@ func TestHammingDistance(t *testing.T) {
 	a := []byte("YELLOW SUBMARINE")
 	dist, err := HammingDistance(a, a)
 	assert.NoError(t, err)
-	assert.Equal(t, dist, uint(0))
+	assert.Equal(t, dist, 0)
 
 	// Different inputs
 	b := []byte("this is a test")
 	c := []byte("wokka wokka!!!")
 	dist, err = HammingDistance(b, c)
 	assert.NoError(t, err)
-	assert.Equal(t, dist, uint(37))
+	assert.Equal(t, dist, 37)
 }
 
 func TestHammingDistanceWithDifferentLengthInputs(t *testing.T) {
